+++ app/scripts/background.js
@@ -14,7 +14,6 @@ const pump = require('pump')
 const debounce = require('debounce-stream')
 const log = require('loglevel')
 const extension = require('extensionizer')
-const LocalStorageStore = require('obs-store/lib/localStorage')
 const LocalStore = require('./lib/local-store')
 const storeTransform = require('obs-store/lib/transform')
 const asStream = require('obs-store/lib/asStream')
@@ -43,7 +42,6 @@ const {
 // METAMASK_TEST_CONFIG is used in e2e tests to set the default network to localhost
 const firstTimeState = Object.assign({}, rawFirstTimeState, global.METAMASK_TEST_CONFIG)
 
-const STORAGE_KEY = 'metamask-config'
 const METAMASK_DEBUG = process.env.METAMASK_DEBUG
 
 log.setDefaultLevel(process.env.METAMASK_DEBUG ? 'debug' : 'warn')
@@ -67,7 +65,6 @@ let notificationIsOpen = false
 const openMetamaskTabsIDs = {}
 
 // state persistence
-const diskStore = new LocalStorageStore({ storageKey: STORAGE_KEY })
 const localStore = new LocalStore()
 let versionedData
 
@@ -75,7 +72,7 @@ let versionedData
 initialize().catch(log.error)
 
 // setup metamask mesh testing container
-setupMetamaskMeshMetrics()
+const { submitMeshMetricsEntry } = setupMetamaskMeshMetrics()
 
 
 /**
@@ -180,7 +177,6 @@ async function loadStateFromPersistence () {
   // read from disk
   // first from preferred, async API:
   versionedData = (await localStore.get()) ||
-                  diskStore.getState() ||
                   migrator.generateInitialState(firstTimeState)
 
   // check if somehow state is empty
@@ -188,21 +184,9 @@ async function loadStateFromPersistence () {
   // for a small number of users
   // https://github.com/metamask/metamask-extension/issues/3919
   if (versionedData && !versionedData.data) {
-    // try to recover from diskStore incase only localStore is bad
-    const diskStoreState = diskStore.getState()
-    if (diskStoreState && diskStoreState.data) {
-      // we were able to recover (though it might be old)
-      versionedData = diskStoreState
-      const vaultStructure = getObjStructure(versionedData)
-      sentry.captureMessage('MetaMask - Empty vault found - recovered from diskStore', {
-        // "extra" key is required by Sentry
-        extra: { vaultStructure },
-      })
-    } else {
-      // unable to recover, clear state
-      versionedData = migrator.generateInitialState(firstTimeState)
-      sentry.captureMessage('MetaMask - Empty vault found - unable to recover')
-    }
+    // unable to recover, clear state
+    versionedData = migrator.generateInitialState(firstTimeState)
+    sentry.captureMessage('MetaMask - Empty vault found - unable to recover')
   }
 
   // report migration errors to sentry
@@ -249,11 +233,13 @@ function setupController (initState, initLangCode) {
   //
   // MetaMask Controller
   //
+  const { ABTestController = {} } = initState
+  const { abTests = {} } = ABTestController
 
   const controller = new MetamaskController({
     // User confirmation callbacks:
     showUnconfirmedMessage: triggerUi,
-    showUnapprovedTx: triggerUi,
+    showUnapprovedTx: abTests.fullScreenVsPopup === 'fullScreen' ? triggerUiInNewTab : triggerUi,
     openPopup: openPopup,
     closePopup: notificationManager.closePopup.bind(notificationManager),
     // initial state
@@ -267,6 +253,11 @@ function setupController (initState, initLangCode) {
 
   const provider = controller.provider
   setupEnsIpfsResolver({ provider })
+
+  // submit rpc requests to mesh-metrics
+  controller.networkController.on('rpc-req', (data) => {
+    submitMeshMetricsEntry({ type: 'rpc', data })
+  })
 
   // report failed transactions to Sentry
   controller.txController.on(`tx:status-update`, (txId, status) => {
@@ -450,6 +441,20 @@ function triggerUi () {
       notificationIsOpen = true
     }
   })
+}
+
+/**
+ * Opens a new browser tab for user confirmation
+ */
+function triggerUiInNewTab () {
+  const tabIdsArray = Object.keys(openMetamaskTabsIDs)
+  if (tabIdsArray.length) {
+    extension.tabs.update(parseInt(tabIdsArray[0], 10), { 'active': true }, () => {
+      extension.tabs.reload(parseInt(tabIdsArray[0], 10))
+    })
+  } else {
+    platform.openExtensionInBrowser()
+  }
 }
 
 /**