+++ app/scripts/controllers/incoming-transactions.js
@@ -7,17 +7,20 @@ import fetchWithTimeout from '../lib/fetch-with-timeout'
 const {
   MAINNET_CODE,
   ROPSTEN_CODE,
-  RINKEYBY_CODE,
+  RINKEBY_CODE,
   KOVAN_CODE,
+  GOERLI_CODE,
   ROPSTEN,
   RINKEBY,
   KOVAN,
+  GOERLI,
   MAINNET,
 } = require('./network/enums')
 const networkTypeToIdMap = {
   [ROPSTEN]: String(ROPSTEN_CODE),
-  [RINKEBY]: String(RINKEYBY_CODE),
+  [RINKEBY]: String(RINKEBY_CODE),
   [KOVAN]: String(KOVAN_CODE),
+  [GOERLI]: String(GOERLI_CODE),
   [MAINNET]: String(MAINNET_CODE),
 }
 const fetch = fetchWithTimeout({
@@ -52,6 +55,7 @@ class IncomingTransactionsController {
         [ROPSTEN]: null,
         [RINKEBY]: null,
         [KOVAN]: null,
+        [GOERLI]: null,
         [MAINNET]: null,
       },
     }, opts.initState)
@@ -171,18 +175,14 @@ class IncomingTransactionsController {
   }
 
   async _fetchAll (address, fromBlock, networkType) {
-    try {
-      const fetchedTxResponse = await this._fetchTxs(address, fromBlock, networkType)
-      return this._processTxFetchResponse(fetchedTxResponse)
-    } catch (err) {
-      log.error(err)
-    }
+    const fetchedTxResponse = await this._fetchTxs(address, fromBlock, networkType)
+    return this._processTxFetchResponse(fetchedTxResponse)
   }
 
   async _fetchTxs (address, fromBlock, networkType) {
     let etherscanSubdomain = 'api'
     const currentNetworkID = networkTypeToIdMap[networkType]
-    const supportedNetworkTypes = [ROPSTEN, RINKEBY, KOVAN, MAINNET]
+    const supportedNetworkTypes = [ROPSTEN, RINKEBY, KOVAN, GOERLI, MAINNET]
 
     if (supportedNetworkTypes.indexOf(networkType) === -1) {
       return {}
@@ -207,7 +207,7 @@ class IncomingTransactionsController {
     }
   }
 
-  _processTxFetchResponse ({ status, result, address, currentNetworkID }) {
+  _processTxFetchResponse ({ status, result = [], address, currentNetworkID }) {
     if (status !== '0' && result.length > 0) {
       const remoteTxList = {}
       const remoteTxs = []