+++ test/setup.js
@@ -5,3 +5,6 @@ require('@babel/register')({
 require('./helper')
 
 window.SVGPathElement = window.SVGPathElement || { prototype: {} }
+window.fetch = window.fetch || function fetch () { return Promise.resolve() }
+global.indexedDB = {}
+global.fetch = global.fetch || function fetch () { return Promise.resolve() }