+++ test/unit/ui/app/actions.spec.js
@@ -55,6 +55,12 @@ describe('Actions', () => {
       initState: clone(firstTimeState),
     })
 
+    metamaskController.threeBoxController = {
+      new3Box: sinon.spy(),
+      getThreeBoxAddress: sinon.spy(),
+      getThreeBoxSyncingState: sinon.spy(),
+    }
+
     await metamaskController.createNewVaultAndRestore(password, TEST_SEED)
 
     await metamaskController.importAccountWithStrategy('Private Key', [ importPrivkey ])