+++ ui/app/pages/confirm-transaction-base/confirm-transaction-base.component.js
@@ -15,6 +15,7 @@ import { CONFIRMED_STATUS, DROPPED_STATUS } from '../../helpers/constants/transa
 import UserPreferencedCurrencyDisplay from '../../components/app/user-preferenced-currency-display'
 import { PRIMARY, SECONDARY } from '../../helpers/constants/common'
 import AdvancedGasInputs from '../../components/app/gas-customization/advanced-gas-inputs'
+import TextField from '../../components/ui/text-field'
 
 export default class ConfirmTransactionBase extends Component {
   static contextTypes = {
@@ -50,6 +51,9 @@ export default class ConfirmTransactionBase extends Component {
     isTxReprice: PropTypes.bool,
     methodData: PropTypes.object,
     nonce: PropTypes.string,
+    useNonceField: PropTypes.bool,
+    customNonceValue: PropTypes.string,
+    updateCustomNonce: PropTypes.func,
     assetImage: PropTypes.string,
     sendTransaction: PropTypes.func,
     showCustomizeGasModal: PropTypes.func,
@@ -96,11 +100,14 @@ export default class ConfirmTransactionBase extends Component {
     insufficientBalance: PropTypes.bool,
     hideFiatConversion: PropTypes.bool,
     transactionCategory: PropTypes.string,
+    getNextNonce: PropTypes.func,
+    nextNonce: PropTypes.number,
   }
 
   state = {
     submitting: false,
     submitError: null,
+    submitWarning: '',
   }
 
   componentDidUpdate (prevProps) {
@@ -109,10 +116,20 @@ export default class ConfirmTransactionBase extends Component {
       showTransactionConfirmedModal,
       history,
       clearConfirmTransaction,
+      nextNonce,
+      customNonceValue,
     } = this.props
     const { transactionStatus: prevTxStatus } = prevProps
     const statusUpdated = transactionStatus !== prevTxStatus
     const txDroppedOrConfirmed = transactionStatus === DROPPED_STATUS || transactionStatus === CONFIRMED_STATUS
+
+    if (nextNonce !== prevProps.nextNonce || customNonceValue !== prevProps.customNonceValue) {
+      if (customNonceValue > nextNonce) {
+        this.setState({ submitWarning: this.context.t('nextNonceWarning', [nextNonce]) })
+      } else {
+        this.setState({ submitWarning: '' })
+      }
+    }
 
     if (statusUpdated && txDroppedOrConfirmed) {
       showTransactionConfirmedModal({
@@ -204,11 +221,16 @@ export default class ConfirmTransactionBase extends Component {
       hexTransactionFee,
       hexTransactionTotal,
       hideDetails,
+      useNonceField,
+      customNonceValue,
+      updateCustomNonce,
       advancedInlineGasShown,
       customGas,
       insufficientBalance,
       updateGasAndCalculate,
       hideFiatConversion,
+      nextNonce,
+      getNextNonce,
     } = this.props
 
     if (hideDetails) {
@@ -240,7 +262,7 @@ export default class ConfirmTransactionBase extends Component {
               : null
             }
           </div>
-          <div>
+          <div className={useNonceField ? 'confirm-page-container-content__gas-fee' : null}>
             <ConfirmDetailRow
               label="Total"
               value={hexTransactionTotal}
@@ -251,6 +273,31 @@ export default class ConfirmTransactionBase extends Component {
               primaryValueTextColor="#2f9ae0"
             />
           </div>
+          {useNonceField ? <div>
+            <div className="confirm-detail-row">
+              <div className="confirm-detail-row__label">
+                { this.context.t('nonceFieldHeading') }
+              </div>
+              <div className="custom-nonce-input">
+                <TextField
+                  type="number"
+                  min="0"
+                  placeholder={ nextNonce ? nextNonce.toString() : null }
+                  onChange={({ target: { value } }) => {
+                    if (!value.length || Number(value) < 0) {
+                      updateCustomNonce('')
+                    } else {
+                      updateCustomNonce(String(Math.floor(value)))
+                    }
+                    getNextNonce()
+                  }}
+                  fullWidth
+                  margin="dense"
+                  value={ customNonceValue || '' }
+                />
+              </div>
+            </div>
+          </div> : null}
         </div>
       )
     )
@@ -347,21 +394,32 @@ export default class ConfirmTransactionBase extends Component {
 
   handleCancel () {
     const { metricsEvent } = this.context
-    const { onCancel, txData, cancelTransaction, history, clearConfirmTransaction, actionKey, txData: { origin }, methodData = {} } = this.props
+    const {
+      onCancel,
+      txData,
+      cancelTransaction,
+      history,
+      clearConfirmTransaction,
+      actionKey,
+      txData: { origin },
+      methodData = {},
+      updateCustomNonce,
+    } = this.props
 
+    metricsEvent({
+      eventOpts: {
+        category: 'Transactions',
+        action: 'Confirm Screen',
+        name: 'Cancel',
+      },
+      customVariables: {
+        recipientKnown: null,
+        functionType: actionKey || getMethodName(methodData.name) || 'contractInteraction',
+        origin,
+      },
+    })
+    updateCustomNonce('')
     if (onCancel) {
-      metricsEvent({
-        eventOpts: {
-          category: 'Transactions',
-          action: 'Confirm Screen',
-          name: 'Cancel',
-        },
-        customVariables: {
-          recipientKnown: null,
-          functionType: actionKey || getMethodName(methodData.name) || 'contractInteraction',
-          origin,
-        },
-      })
       onCancel(txData)
     } else {
       cancelTransaction(txData)
@@ -374,7 +432,19 @@ export default class ConfirmTransactionBase extends Component {
 
   handleSubmit () {
     const { metricsEvent } = this.context
-    const { txData: { origin }, sendTransaction, clearConfirmTransaction, txData, history, onSubmit, actionKey, metaMetricsSendCount = 0, setMetaMetricsSendCount, methodData = {} } = this.props
+    const {
+      txData: { origin },
+      sendTransaction,
+      clearConfirmTransaction,
+      txData,
+      history,
+      onSubmit,
+      actionKey,
+      metaMetricsSendCount = 0,
+      setMetaMetricsSendCount,
+      methodData = {},
+      updateCustomNonce,
+    } = this.props
     const { submitting } = this.state
 
     if (submitting) {
@@ -406,6 +476,7 @@ export default class ConfirmTransactionBase extends Component {
                 this.setState({
                   submitting: false,
                 })
+                updateCustomNonce('')
               })
           } else {
             sendTransaction(txData)
@@ -415,6 +486,7 @@ export default class ConfirmTransactionBase extends Component {
                   submitting: false,
                 }, () => {
                   history.push(DEFAULT_ROUTE)
+                  updateCustomNonce('')
                 })
               })
               .catch(error => {
@@ -422,6 +494,7 @@ export default class ConfirmTransactionBase extends Component {
                   submitting: false,
                   submitError: error.message,
                 })
+                updateCustomNonce('')
               })
           }
         })
@@ -493,7 +566,7 @@ export default class ConfirmTransactionBase extends Component {
   }
 
   componentDidMount () {
-    const { txData: { origin, id } = {}, cancelTransaction } = this.props
+    const { txData: { origin, id } = {}, cancelTransaction, getNextNonce } = this.props
     const { metricsEvent } = this.context
     metricsEvent({
       eventOpts: {
@@ -521,6 +594,8 @@ export default class ConfirmTransactionBase extends Component {
         cancelTransaction({ id })
       }
     }
+
+    getNextNonce()
   }
 
   render () {
@@ -543,12 +618,13 @@ export default class ConfirmTransactionBase extends Component {
       contentComponent,
       onEdit,
       nonce,
+      customNonceValue,
       assetImage,
       warning,
       unapprovedTxCount,
       transactionCategory,
     } = this.props
-    const { submitting, submitError } = this.state
+    const { submitting, submitError, submitWarning } = this.state
 
     const { name } = methodData
     const { valid, errorKey } = this.getErrorKey()
@@ -572,13 +648,13 @@ export default class ConfirmTransactionBase extends Component {
         detailsComponent={this.renderDetails()}
         dataComponent={this.renderData()}
         contentComponent={contentComponent}
-        nonce={nonce}
+        nonce={customNonceValue || nonce}
         unapprovedTxCount={unapprovedTxCount}
         assetImage={assetImage}
         identiconAddress={identiconAddress}
         errorMessage={errorMessage || submitError}
         errorKey={propsErrorKey || errorKey}
-        warning={warning}
+        warning={warning || submitWarning}
         totalTx={totalTx}
         positionOfCurrentTx={positionOfCurrentTx}
         nextTxId={nextTxId}