+++ ui/app/pages/send/send-content/send-row-wrapper/send-row-wrapper.component.js
@@ -1,7 +1,6 @@
 import React, { Component } from 'react'
 import PropTypes from 'prop-types'
 import SendRowErrorMessage from './send-row-error-message'
-import SendRowWarningMessage from './send-row-warning-message'
 
 export default class SendRowWrapper extends Component {
 
@@ -10,8 +9,6 @@ export default class SendRowWrapper extends Component {
     errorType: PropTypes.string,
     label: PropTypes.string,
     showError: PropTypes.bool,
-    showWarning: PropTypes.bool,
-    warningType: PropTypes.string,
   };
 
   static contextTypes = {
@@ -24,8 +21,6 @@ export default class SendRowWrapper extends Component {
       errorType = '',
       label,
       showError = false,
-      showWarning = false,
-      warningType = '',
     } = this.props
     const formField = Array.isArray(children) ? children[1] || children[0] : children
     const customLabelContent = children.length > 1 ? children[0] : null
@@ -42,7 +37,6 @@ export default class SendRowWrapper extends Component {
           </div>
           <div>
             {showError && <SendRowErrorMessage errorType={errorType} />}
-            {!showError && showWarning && <SendRowWarningMessage warningType={warningType} />}
           </div>
         </div>
       </div>
@@ -55,8 +49,6 @@ export default class SendRowWrapper extends Component {
       errorType = '',
       label,
       showError = false,
-      showWarning = false,
-      warningType = '',
     } = this.props
 
     const formField = Array.isArray(children) ? children[1] || children[0] : children
@@ -67,7 +59,6 @@ export default class SendRowWrapper extends Component {
         <div className="send-v2__form-label">
           {label}
           {showError && <SendRowErrorMessage errorType={errorType} />}
-          {!showError && showWarning && <SendRowWarningMessage warningType={warningType} />}
           {customLabelContent}
         </div>
         <div className="send-v2__form-field">