+++ ui/app/pages/send/send.selectors.js
@@ -53,7 +53,6 @@ const selectors = {
   getSendTo,
   getSendToAccounts,
   getSendToNickname,
-  getSendWarnings,
   getTokenBalance,
   getTokenExchangeRate,
   getUnapprovedTxs,
@@ -244,11 +243,6 @@ function getSendToAccounts (state) {
   const addressBookAccounts = getAddressBook(state)
   return [...fromAccounts, ...addressBookAccounts]
 }
-
-function getSendWarnings (state) {
-  return state.send.warnings
-}
-
 function getTokenBalance (state) {
   return state.metamask.send.tokenBalance
 }